backend/requirements.txt
fastapi
uvicorn[standard]
redis
ccxt
websockets
sqlalchemy
asyncpg
alembic
pydantic-settings
httpx
pgvector
numpy
pandas==2.2.1
